SYS-1704 - Reorganzie script
