+++ app/attributes/dns_record_value
@@ -1,0 +1,1 @@
+eph-666666-pseudonymised-instance-1598469260.clt2hqjzc0dv.eu-west-1.rds.amazonaws.com
